testing/fake_data/pyproject.toml
[tool.poetry]
name = "fake_data"
version = "0.1.0"
description = ""
authors = ["tuilagio <ba.q.le@campus.tu-berlin.de>"]

[tool.poetry.dependencies]
python = "^3.9"
psycopg2 = "^2.9.1"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
aiosql = "^3.3.1"
autopep8 = "^1.6.0"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
